% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ipstack-package.R
\docType{package}
\name{ipstack}
\alias{ipstack}
\alias{ipstack-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/ipstack}
\item BugReports: \url{https://gitlab.com/hrbrmstr/ipstack/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
